+++ db/app/admin/src/full/R__18.OBJECT_GRANT.ADMIN_SCHEDULER_PKG.sql
@@ -1,0 +1,4 @@
+call dbms_application_info.set_module('R__18.OBJECT_GRANT.ADMIN_SCHEDULER_PKG.sql', null);
+call dbms_application_info.set_action('SQL statement 1');
+GRANT EXECUTE ON "ADMIN_SCHEDULER_PKG" TO PUBLIC;
+